New module for the purge_dups program to remove haplotigs from assemblies.
I implemented all the steps in a single process.  Most commands run
quickly on a single core, but the self-alignment takes advantage of
multiple CPUs.  The purge_dups container ships its own copy of
minimap2, and it is simpler to just use it for the genome self-alignment.
Nevertheless, the process expects to receive the reads mapped on
the genome by a separate process.  This is useful because this read
mapping is also needed for other modules such as racon.

The `lambda_top10_overlaps.paf` file needed to be replaced by a fresh
copy of the minimap2 module's test output, otherwise one of the commands
of purge_dups commands would crash on it.