Split the last_filter_maf process in two for one-to-one and one-to-many alns.
The `last_align module` produces a many-to-many alignment.  One unique feature
of _LAST_ is the possibility to reduce it to a one-to-many or a one-to-one
alignment with the `last-split` tool.  The current `last_filter_maf` process
filtrers to a one-to-one alignment.  This branch splits and renames the
module in two: `last_filter_one_to_one` and `last_filter_one_to_many`.

The process contain either a `last-split` command (last_filter_one_to_many)
or a pipe of `last-split` and `maf-swap` commands.  Modularising further
in smaller processes would be wasteful as the alignment files are large
and the intermediate steps of the pipe do not need to be stored.

The `last_filter_one_to_many` process does not produce a `.tab` file because
I am not aware of a use case for it.  (It is usually consumed for producing
dot plots of one-to-one or many-to-many alignments.)